Add HappyHTTP library.
We need a http communication library for license requests. HappyHTTP
is lightweight and cross platform. Compared to the url fetcher in
Chromium, it is simpler and easier to use, and it is much smaller.

The source is cloned from https://github.com/Zintinio/HappyHTTP/
SHA 7306b1606a09063ac38c264afe59f0ad0b441750.

Change-Id: I21b691452f3d7b5fea34c611012b252e2e3c9e15