[doc] Fix broken link in DockerHub
While not ideal, absolute URLs are used as it is the simplest way to
make the links work on GitHub and Docker Hub. These links in cloned
repositories will point back to the main repository and if it is an
issue, the links in the cloned repository need to be updated.

Fixes #408.

Change-Id: I6d9e4bf258325afaefa6c600daf973d0dcf5a94d