build: Fix builds with CC=clang CXX=clang++
When using CC=clang CXX=clang++, there is a binutils version check
that does not work correctly in common.gypi.  Since we are stuck with
a very old version of chromium/src/build, there is nothing to do but
patch it to remove the check.  Thankfully, this version number does
not control anything critical in the build settings as far as we can
tell.

Change-Id: Id749d97c5898917592f66136538ee0fa5ca78767