build: Make release binary names more consistent
Now that we have multiple architectures, we should factor both OS and
architecture into the names of release binaries.  This makes the names
more formulaic, as well as consistent with the static-ffmpeg-binaries
repository.  Shaka Streamer will pull binaries from both this repo and
that one, so consistent names would be helpful.

The pssh-box release is actually OS and architecture independent, so
remove the suffix from that and only release one copy of it.

Change-Id: Ief3de49fae267c5267647a8dd4377023777ead37